refactor(slc-spacy): Extracted some modules
