+++ duui-slc-spacy/justfile
@@ -1,4 +1,4 @@
-export VERSION := "0.0.2"
+export VERSION := "0.0.3"
 export REPOSITORY := "docker.texttechnologylab.org"
 export BASE_NAME := "duui-slc-spacy"
