+++ duui-slc-spacy/src/main/python/app/__init__.py
@@ -1,0 +1,6 @@
+import logging
+
+logging.basicConfig(
+    level=logging.INFO,
+    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+)
